+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Maintainers:
 
 - NO CHANGELOG AVAILABLE
 
-## September 04, 2023
+## September 05, 2023
 
 - [x] Project structure set up.
 - [x] Following files added:
@@ -31,6 +31,13 @@ Maintainers:
     `README.md`
     `TODO.md`
     `CHANGELOG.md`
-
+    
+## September 06, 2023
+- [x] Added `apiserver/app.py` to create Flask App.
+- [x] Added	`apiserver/config.py` to add default configuration to Flask App.
+- [x] Added	`apiserver/extensions.py` for extension(s) registration.
+- [x] Added	`apiserver/manage.py` to support CLI management of Flask App
+- [x] Added	`apiserver/run.py` to define entry point.
+- [x] `CHANGELOG.md` updated.
 
 ---